requirements.txt
streamlit==0.75.0
glob2==0.7
Pillow==8.1.0
numpy==1.20.0
opencv-python==4.2.0.34